+++ examples/multilingual.rs
@@ -20,7 +20,8 @@ fn get_labels_german() -> FileDialogLabels {
         overwrite: "Überschreiben".to_string(),
 
         reload: "⟲  Neu laden".to_string(),
-        working_directory: "Arbeitsverzeichnis öffnen".to_string(),
+        working_directory: "↗  Arbeitsverzeichnis öffnen".to_string(),
+        select_all: "🗐  Alle auswählen".to_string(),
         show_hidden: " Versteckte Dateien anzeigen".to_string(),
         show_system_files: " Systemdateien anzeigen".to_string(),